refactor: moving the application folder inside the core and updating the imports
